feat: add dto model (#38)
