+++ iOS/issue-tracker/issue-tracker/Supporting Files/Assets.xcassets/vector.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "Vector.png",
+      "filename" : "vector.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
